@@ -1,8 +1,8 @@
-<?php declare(strict_types=1);
+<?php
+declare(strict_types=1);
 
 namespace PcComponentes\OpenApiMessagingContext\OpenApi;
 
 final class JsonValidationException extends \Exception
 {
-
 }
